@TheEpicDolphin 2025-07-18
The tree attention backends supports validating tree drafts, and the eagle proposer can also propose tree drafts, but right now the eagle propser will only generate chains because tree scoring is not yet implemented. Scoring is the final missing piece until we can allow the eagle proposer to generate tree drafts.