@TheEpicDolphin 2025-07-18
@wangjiahe0915 this tree attention backend supports both chain and tree attention masks. The Eagle proposer also now supports generating tree drafts. However, tree scoring has still not been implemented, and so we cannot yet enable tree drafting.

If you were to attempt to generate tree drafts on this PR, the system would score the drafts as if they were a chain. So you can't yet benefit from a tree draft. Tree scoring is the last missing piece that I will work on next