@sgrigory 2025-07-24
> @houseroad Correct, this tree attention backend is for speculative decoding to enable drafting, validating, and scoring tree drafts. The benchmark data is in the summary of this PR. It compares FA3 (before this PR), FA3 with this PR (no change), and tree attention. I still need to address the performance gap but I will iterate in the coming weeks to improve. Additionally, tree scoring is not yet implemented so the benchmarks do not reflect the benefits from tree drafts.
> 
> According to Grigory, the perf gap can most likely be attributed to the forward attention methods not yet being cuda graphed. Doing so would greatly reduce the CPU overhead from the separate prefix and suffix attention kernel launching, and also the attention merge operation!

Yes, we optimised tree attention for CUDA-graph mode. If you call it in eager, there is a lot of CPU overhead. This will be eliminated once https://github.com/vllm-project/vllm/pull/21477 lands

Modifying unified triton attention looks like a very nice approach. Tree attention was developed to avoid loading an additional tensor with an explicit mask. Without modifying the existing kernels and for variable seqlens, this mask would have been (batch_size, context_length, draft_length). However, for you the mask is of shape `(draft_length, draft_length)`, which is tiny, so its loading is fast