@LucasWilkinson 2025-07-30
> @LucasWilkinson please see the test plan for the benchmark of Eagle with no tree attention (flash attention) comparing this PR and main. There is no regression

sorry I missed this was with eagle on; thanks!