@wangjiahe0915 2025-07-30
I have identified two abnormal cases:

1. When a single sequence length is short and less than tree_attn_bias[0], during the prefill stage, decode_metadata is used, and it is incorrect for tree_attn_bias to participate in the calculation at this time.

2. When the generated eagle draft sequences are executed together with newly added scheduled sequences in the main model's tree_attention, decode_metadata is used. In this case, does tree_attn_bias only apply to the eagle draft sequences? For the newly added sequences, decode_metadata is still used—will tree_attn_bias be used in unified_attention for these sequences?