@TheEpicDolphin 2025-07-30
@wangjiahe0915 

1. Correct, for prefill sequences that happen to be smaller than the tree size, it will be incorrectly masked. I think that this is just a limitation of how v1 currently distinguishes between prefill and decode. This definitely needs refactoring, and probably already is in the works.

2. I am not 100% sure what is meant by “scheduled sequences”. Is that the token generated by the base model? Either way, this tree attention backend ensures that the base model only applies tree masking on the eagle drafted tokens. The rest of the sequence (prompt + accepted tokens so far) just have normal masking