@wangjiahe0915 2025-07-31
I have another question. The tree structure you currently provide is fixed, which I understand follows the idea of eagle1. When the width and depth increase, this static tree may result in more blocks being allocated in vLLM. If a dynamic tree is used, the tree_attn_mask structure constructed during initialization would no longer be applicable. Is the reason for your design that the efficiency of a dynamic tree would be lower than that of a static tree?